auth

+++ server/routes/student.py
@@ -1,6 +1,7 @@
 from flask import jsonify, request
 from server import app
 from server.models.student import Student
+from flask_login import current_user, login_required
 
 
 @app.route('/api/students', methods=['POST'])
@@ -28,8 +29,12 @@ def create_student():
 
 # Add starred post to student profile
 @app.route('/api/students/<string:net_id>/<int:post_id>', methods=['POST'])
+@login_required
 def add_favorited_project(net_id, post_id):
     """ Add a starred post to a student profile. """
+    if not current_user.net_id == net_id:
+        return jsonify({"status": "error"})
+
     if Student.add_favorited_project(net_id, post_id):
         return jsonify({"status": "success"})
     else:
@@ -37,8 +42,12 @@ def add_favorited_project(net_id, post_id):
 
 
 @app.route('/api/students/<string:net_id>/<int:post_id>', methods=['DELETE'])
+@login_required
 def delete_favorited_project(net_id, post_id):
     """ Remove a starred post from a student profile. """
+    if not current_user.net_id == net_id:
+        return jsonify({"status": "error"})
+
     if Student.delete_favorited_project(net_id, post_id):
         return jsonify({"status": "success"})
     else:
